backend/requirements.txt
# Flask
flask==2.3.3
flask-cors==4.0.0

# LLM
annotated-types==0.7.0
anyio==4.9.0
anthropic>=0.18.0
asttokens==3.0.0
certifi==2025.6.15
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.90.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pywin32==310
pyzmq==27.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
wcwidth==0.2.13
google-generativeai>=0.3.0
